        \name{CCMnet-internal} 
        \alias{NS_Multinomial_uni} 
        \alias{uni_modal_constr} 
        \alias{bi_modal_constr}  
        \title{Internal CCMnet Functions} 
        \description{ 
          Internal CCMnet functions 
        } 
        \details{ 
          These are not to be called by the user. 
        } 
        \keyword{ internal } 